chore: refactor cookie/auth code
